move AlertRuleGroupWithFolderTitle to ngalert models and adapter functions to api's compat
