+++ pkg/services/ngalert/api/compat.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"encoding/json"
 	"time"
 
 	"github.com/prometheus/common/model"
@@ -124,4 +125,93 @@ func ApiAlertRuleGroupFromAlertRuleGroup(d models.AlertRuleGroup) definitions.Al
 		Interval:  d.Interval,
 		Rules:     rules,
 	}
+}
+
+// AlertingFileExportFromAlertRuleGroupWithFolderTitle creates an definitions.AlertingFileExport DTO from []models.AlertRuleGroupWithFolderTitle.
+func AlertingFileExportFromAlertRuleGroupWithFolderTitle(groups []models.AlertRuleGroupWithFolderTitle) (definitions.AlertingFileExport, error) {
+	f := definitions.AlertingFileExport{APIVersion: 1}
+	for _, group := range groups {
+		export, err := AlertRuleGroupExportFromAlertRuleGroupWithFolderTitle(group)
+		if err != nil {
+			return definitions.AlertingFileExport{}, err
+		}
+		f.Groups = append(f.Groups, export)
+	}
+	return f, nil
+}
+
+// AlertRuleGroupExportFromAlertRuleGroupWithFolderTitle creates a definitions.AlertRuleGroupExport DTO from models.AlertRuleGroup.
+func AlertRuleGroupExportFromAlertRuleGroupWithFolderTitle(d models.AlertRuleGroupWithFolderTitle) (definitions.AlertRuleGroupExport, error) {
+	rules := make([]definitions.AlertRuleExport, 0, len(d.Rules))
+	for i := range d.Rules {
+		alert, err := AlertRuleExportFromAlertRule(d.Rules[i])
+		if err != nil {
+			return definitions.AlertRuleGroupExport{}, err
+		}
+		rules = append(rules, alert)
+	}
+	return definitions.AlertRuleGroupExport{
+		OrgID:    d.OrgID,
+		Name:     d.Title,
+		Folder:   d.FolderTitle,
+		Interval: model.Duration(time.Duration(d.Interval) * time.Second),
+		Rules:    rules,
+	}, nil
+}
+
+// AlertRuleExportFromAlertRule creates a definitions.AlertRuleExport DTO from models.AlertRule.
+func AlertRuleExportFromAlertRule(rule models.AlertRule) (definitions.AlertRuleExport, error) {
+	data := make([]definitions.AlertQueryExport, 0, len(rule.Data))
+	for i := range rule.Data {
+		query, err := AlertQueryExportFromAlertQuery(rule.Data[i])
+		if err != nil {
+			return definitions.AlertRuleExport{}, err
+		}
+		data = append(data, query)
+	}
+
+	var dashboardUID string
+	if rule.DashboardUID != nil {
+		dashboardUID = *rule.DashboardUID
+	}
+
+	var panelID int64
+	if rule.PanelID != nil {
+		panelID = *rule.PanelID
+	}
+
+	return definitions.AlertRuleExport{
+		UID:          rule.UID,
+		Title:        rule.Title,
+		For:          model.Duration(rule.For),
+		Condition:    rule.Condition,
+		Data:         data,
+		DashboardUID: dashboardUID,
+		PanelID:      panelID,
+		NoDataState:  definitions.NoDataState(rule.NoDataState),
+		ExecErrState: definitions.ExecutionErrorState(rule.ExecErrState),
+		Annotations:  rule.Annotations,
+		Labels:       rule.Labels,
+		IsPaused:     rule.IsPaused,
+	}, nil
+}
+
+// AlertQueryExportFromAlertQuery creates a definitions.AlertQueryExport DTO from models.AlertQuery.
+func AlertQueryExportFromAlertQuery(query models.AlertQuery) (definitions.AlertQueryExport, error) {
+	// We unmarshal the json.RawMessage model into a map in order to facilitate yaml marshalling.
+	var mdl map[string]interface{}
+	err := json.Unmarshal(query.Model, &mdl)
+	if err != nil {
+		return definitions.AlertQueryExport{}, err
+	}
+	return definitions.AlertQueryExport{
+		RefID:     query.RefID,
+		QueryType: query.QueryType,
+		RelativeTimeRange: definitions.RelativeTimeRange{
+			From: definitions.Duration(query.RelativeTimeRange.From),
+			To:   definitions.Duration(query.RelativeTimeRange.To),
+		},
+		DatasourceUID: query.DatasourceUID,
+		Model:         mdl,
+	}, nil
 }
