+++ pkg/services/ngalert/provisioning/alert_rules.go
@@ -11,8 +11,6 @@ import (
 	"github.com/grafana/grafana/pkg/services/dashboards"
 	"github.com/grafana/grafana/pkg/services/ngalert/models"
 	"github.com/grafana/grafana/pkg/services/ngalert/store"
-	"github.com/grafana/grafana/pkg/services/provisioning/alerting/file"
-
 	"github.com/grafana/grafana/pkg/services/quota"
 	"github.com/grafana/grafana/pkg/util"
 )
@@ -406,17 +404,17 @@ func (service *AlertRuleService) deleteRules(ctx context.Context, orgID int64, t
 }
 
 // GetAlertRuleGroupWithFolderTitle returns the alert rule group with folder title.
-func (service *AlertRuleService) GetAlertRuleGroupWithFolderTitle(ctx context.Context, orgID int64, namespaceUID, group string) (file.AlertRuleGroupWithFolderTitle, error) {
+func (service *AlertRuleService) GetAlertRuleGroupWithFolderTitle(ctx context.Context, orgID int64, namespaceUID, group string) (models.AlertRuleGroupWithFolderTitle, error) {
 	q := models.ListAlertRulesQuery{
 		OrgID:         orgID,
 		NamespaceUIDs: []string{namespaceUID},
 		RuleGroup:     group,
 	}
 	if err := service.ruleStore.ListAlertRules(ctx, &q); err != nil {
-		return file.AlertRuleGroupWithFolderTitle{}, err
+		return models.AlertRuleGroupWithFolderTitle{}, err
 	}
 	if len(q.Result) == 0 {
-		return file.AlertRuleGroupWithFolderTitle{}, store.ErrAlertRuleGroupNotFound
+		return models.AlertRuleGroupWithFolderTitle{}, store.ErrAlertRuleGroupNotFound
 	}
 
 	dq := dashboards.GetDashboardQuery{
@@ -425,10 +423,10 @@ func (service *AlertRuleService) GetAlertRuleGroupWithFolderTitle(ctx context.Co
 	}
 	dash, err := service.dashboardService.GetDashboard(ctx, &dq)
 	if err != nil {
-		return file.AlertRuleGroupWithFolderTitle{}, err
+		return models.AlertRuleGroupWithFolderTitle{}, err
 	}
 
-	res := file.AlertRuleGroupWithFolderTitle{
+	res := models.AlertRuleGroupWithFolderTitle{
 		AlertRuleGroup: &models.AlertRuleGroup{
 			Title:     q.Result[0].RuleGroup,
 			FolderUID: q.Result[0].NamespaceUID,
@@ -447,7 +445,7 @@ func (service *AlertRuleService) GetAlertRuleGroupWithFolderTitle(ctx context.Co
 }
 
 // GetAlertGroupsWithFolderTitle returns all groups with folder title that have at least one alert.
-func (service *AlertRuleService) GetAlertGroupsWithFolderTitle(ctx context.Context, orgID int64) ([]file.AlertRuleGroupWithFolderTitle, error) {
+func (service *AlertRuleService) GetAlertGroupsWithFolderTitle(ctx context.Context, orgID int64) ([]models.AlertRuleGroupWithFolderTitle, error) {
 	q := models.ListAlertRulesQuery{
 		OrgID: orgID,
 	}
@@ -484,13 +482,13 @@ func (service *AlertRuleService) GetAlertGroupsWithFolderTitle(ctx context.Conte
 		folderUidToTitle[dash.UID] = dash.Title
 	}
 
-	result := make([]file.AlertRuleGroupWithFolderTitle, 0)
+	result := make([]models.AlertRuleGroupWithFolderTitle, 0)
 	for groupKey, rules := range groups {
 		title, ok := folderUidToTitle[groupKey.NamespaceUID]
 		if !ok {
 			return nil, fmt.Errorf("cannot find title for folder with uid '%s'", groupKey.NamespaceUID)
 		}
-		result = append(result, file.AlertRuleGroupWithFolderTitle{
+		result = append(result, models.AlertRuleGroupWithFolderTitle{
 			AlertRuleGroup: &models.AlertRuleGroup{
 				Title:     rules[0].RuleGroup,
 				FolderUID: rules[0].NamespaceUID,
